Fix this warning:
HOSTCC   build/romcc
SerialICE/util/romcc.c: In function 'scc_transform':
SerialICE/util/romcc.c:21097: warning: format '%5ld' expects type 'long int', but argument 3 has type 'int'

I hope it's the right fix.





git-svn-id: svn://svn.coreboot.org/serialice/trunk@69 8d503c2a-b2ab-11de-a960-3b3a9759a4e4

+++ SerialICE/util/romcc.c
@@ -21092,7 +21092,7 @@ static void scc_transform(struct compile_state *state)
 
 			if (state->compiler->debug & DEBUG_SCC_TRANSFORM) {
 				fprintf(state->errout, "sedge: %5ld (%5d -> %5d)\n",
-					sedge - scc.ssa_edges,
+					(unsigned long)sedge - (unsigned long)scc.ssa_edges,
 					sedge->src->def->id,
 					sedge->dst->def->id);
 			}
